Added singleton for services.
