+++ EA.DesktopApp/View/RedactorForm.xaml
@@ -8,7 +8,7 @@
         Title="Employees Accounting" Height="400" Width="535" MinHeight="400" MaxHeight="500" MinWidth="535"
         MaxWidth="800">
     <Window.Resources>
-        <viewModels:RedactorFormViewModel x:Key="RedactorFormViewModel" />
+        <viewModels:RedactorViewModel x:Key="RedactorFormViewModel" />
 
         <!-- IData Error style  -->
