+++ EA.DesktopApp/ViewModels/AdminViewModel.cs
@@ -1,46 +1,35 @@
 ﻿using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
-using System.Windows;
 using System.Windows.Input;
 using EA.DesktopApp.Services;
-using EA.DesktopApp.View;
 using EA.DesktopApp.ViewModels.Commands;
 
 namespace EA.DesktopApp.ViewModels
 {
-    public class LoginFormViewModel : BaseViewModel, IDataErrorInfo
+    internal class AdminViewModel : BaseViewModel, IDataErrorInfo
     {
-        private readonly LoginWindow _loginWindow;
         private bool _isReady;
         private bool _isRunning;
-        private RegistrationForm _registrationFormPage;
-        private SoundPlayerService _soundPlayerHelper = new SoundPlayerService();
 
         private string _loginValue;
+        private string _oldPasswordValue;
 
         private string _passwordValue;
+        private string _userMessage;
 
-
-        public LoginFormViewModel()
+        public AdminViewModel()
         {
             InitializeCommands();
         }
 
-        public LoginFormViewModel(LoginWindow loginWindow)
-        {
-            this._loginWindow = loginWindow;
-            InitializeCommands();
-        }
+        public string Registration => "Нажмите для регистрации";
+        public string ClearFields => "Нажмите для очистки полей";
 
-        public ICommand LoginCommand { get; private set; }
-        public ICommand CancelCommand { get; private set; }
-        public ICommand AdminModeCommand { get; private set; }
-
-        public string Login { get; } = "Enter the password";
-        public string Cancel { get; } = "Press for clear";
+        public ICommand ClearFieldsCommand { get; set; }
+        public ICommand RegistrationCommand { get; set; }
 
         /// <summary>
-        ///     Start webCam service button toggle
+        ///     Start webCam service button toogle
         /// </summary>
         public bool IsRunning
         {
@@ -61,6 +50,16 @@ namespace EA.DesktopApp.ViewModels
             set
             {
                 _isReady = value;
+                OnPropertyChanged();
+            }
+        }
+
+        public string UserMessage
+        {
+            get => _userMessage;
+            set
+            {
+                _userMessage = value;
                 OnPropertyChanged();
             }
         }
@@ -87,6 +86,17 @@ namespace EA.DesktopApp.ViewModels
             }
         }
 
+        [Required(AllowEmptyStrings = false)]
+        public string OldPasswordField
+        {
+            get => _oldPasswordValue;
+            set
+            {
+                _oldPasswordValue = value;
+                OnPropertyChanged();
+            }
+        }
+
         /// <summary>
         ///     Error indexer
         /// </summary>
@@ -107,7 +117,7 @@ namespace EA.DesktopApp.ViewModels
                         }
                         else if (LoginField.Contains(" "))
                         {
-                            error = "Пароль не может содержать пробел";
+                            error = "Логин не может содержать пробел";
                         }
 
                         break;
@@ -118,6 +128,17 @@ namespace EA.DesktopApp.ViewModels
                             error = "Введите пароль!";
                         }
                         else if (PasswordField.Contains(" "))
+                        {
+                            error = "Пароль не может содержать пробел";
+                        }
+
+                        break;
+                    case "OldPasswordField":
+                        if (string.IsNullOrEmpty(PasswordField))
+                        {
+                            error = "Введите старый пароль!";
+                        }
+                        else if (OldPasswordField.Contains(" "))
                         {
                             error = "Пароль не может содержать пробел";
                         }
@@ -136,67 +157,22 @@ namespace EA.DesktopApp.ViewModels
 
         private void InitializeCommands()
         {
-            LoginCommand = new RelayCommand(ToggleLoginExecute);
-            CancelCommand = new RelayCommand(ToggleCancelExecute);
-            AdminModeCommand = new RelayCommand(ToggleAdminWindowShowExecute);
+            ClearFieldsCommand = new RelayCommand(ToggleClearFieldsExecute);
+            RegistrationCommand = new RelayCommand(ToggleRegistrationExecute);
         }
 
-        private void ToggleLoginExecute()
+        private void ToggleRegistrationExecute()
         {
-            // Playing sound effect for button
-            _soundPlayerHelper = new SoundPlayerService();
-            _soundPlayerHelper.PlaySound("button");
-
-            // True - button is pushed - Working!
-            IsRunning = false;
-            if (_registrationFormPage != null && !_registrationFormPage.IsClosed)
-            {
-                return;
-            }
-
-            var registrationFormViewModel = new RegistrationFormViewModel();
-
-            _registrationFormPage = new RegistrationForm(IsRunning)
-            {
-                DataContext = registrationFormViewModel,
-                Owner = Application.Current.MainWindow
-            };
-
-            //_registrationFormPage.Show();
-            //IsStreaming = false;
-            //_faceDetectionService.CancelServiceAsync();
-            _registrationFormPage.ShowDialog();
-
-            //if (!_faceDetectionService.IsRunning)
-            //{
-            //    IsStreaming = true;
-            //    _faceDetectionService.RunServiceAsync();
-            //}
         }
 
-        private void ToggleCancelExecute()
+        private void ToggleClearFieldsExecute()
         {
-            _soundPlayerHelper = new SoundPlayerService();
-            _soundPlayerHelper.PlaySound("button");
+            var soundHelper = new SoundPlayerService();
+            soundHelper.PlaySound("button");
 
             LoginField = string.Empty;
             PasswordField = string.Empty;
-        }
-
-        private void ToggleAdminWindowShowExecute()
-        {
-            _soundPlayerHelper = new SoundPlayerService();
-            _soundPlayerHelper.PlaySound("button");
-            var adminForm = new AdminForm
-            {
-                Owner = Application.Current.MainWindow
-            };
-            adminForm.ShowDialog();
-        }
-
-        public void ShowWindow()
-        {
-            _loginWindow.ShowDialog();
+            OldPasswordField = string.Empty;
         }
     }
 }
